lesson-2: moved retrofit from RetrofitRepositoriesRepo.kt to App.kt
